template rename to kitties
