+++ projects/kitties/runtime/src/kitties/mod.rs
@@ -1,0 +1,53 @@
+/// Cat Module for runtime
+
+use support::{decl_module, decl_storage, decl_event, StorageValue, dispatch::Result};
+use system::ensure_signed;
+
+/// The module's configuration trait.
+pub trait Trait: system::Trait {
+	// TODO: Add other types and constants required configure this module.
+
+	/// The overarching event type.
+	type Event: From<Event<Self>> + Into<<Self as system::Trait>::Event>;
+}
+
+// This module's storage items.
+decl_storage! {
+	trait Store for Module<T: Trait> as KittyStorage {
+		// Declare storage and getter functions here
+	}
+}
+
+// The module's dispatchable functions.
+decl_module! {
+	/// The module declaration.
+	pub struct Module<T: Trait> for enum Call where origin: T::Origin {
+		// Initializing events
+		// this is needed only if you are using events in your module
+		// fn deposit_event() = default;
+
+		// Just a dummy entry point.
+		// function that can be called by the external world as an extrinsics call
+		// takes a parameter of the type `AccountId`, stores it and emits an event
+		pub fn do_something(origin, something: u32) -> Result {
+			let sender = ensure_signed(origin)?;
+
+			// here we are raising the Something event
+			// Self::deposit_event(RawEvent::SomethingStored(something, who));
+			Ok(())
+		}
+	}
+}
+
+decl_event!(
+	pub enum Event<T> where AccountId = <T as system::Trait>::AccountId {
+		// Just a dummy event.
+		// Event `Something` is declared with a parameter of the type `u32` and `AccountId`
+		// To emit this event, we call the deposit funtion, from our runtime funtions
+		SomethingStored(u32, AccountId),
+	}
+);
+
+// for Module test
+mod mock;
+mod tests;
